+++ app/controllers/komet_dashboard_controller.rb
@@ -291,7 +291,8 @@ class KometDashboardController < ApplicationController
                                   'data-menu-uuid' => raw_node[:id],
                                   'data-menu-state' => raw_node[:state],
                                   'data-menu-concept-text' => raw_node[:text],
-                                  'data-menu-concept-terminology-type' => raw_node[:terminology_type]
+                                  'data-menu-concept-terminology-type' => raw_node[:terminology_type],
+                                  'aria-label' => raw_node[:text]
             }
             parent_search = parent_search_param
             parent_reversed = parent_reversed_param